+++ src/api.rs
@@ -1,4 +1,5 @@
-use crate::api_common::{ApiErr, Result};
+pub use crate::api_common::RecryptErr;
+use crate::api_common::Result;
 use crate::internal;
 use crate::internal::bytedecoder::{BytesDecoder, DecodeErr};
 use crate::internal::curve;
@@ -26,7 +27,7 @@ use std::fmt;
 
 /// Recrypt public API - 256-bit
 #[derive(Debug)]
-pub struct Api<H, S, R> {
+pub struct Recrypt<H, S, R> {
     random_bytes: R,
     sha_256: H,
     ed25519: S,
@@ -35,24 +36,24 @@ pub struct Api<H, S, R> {
     schnorr_signing: SchnorrSign<Monty256, Fr256, H>,
 }
 
-impl Api<Sha256, Ed25519, RandomBytes<rand::rngs::ThreadRng>> {
-    pub fn new() -> Api<Sha256, Ed25519, RandomBytes<rand::rngs::ThreadRng>> {
-        Api::new_with_rand(rand::thread_rng())
+impl Recrypt<Sha256, Ed25519, RandomBytes<rand::rngs::ThreadRng>> {
+    pub fn new() -> Recrypt<Sha256, Ed25519, RandomBytes<rand::rngs::ThreadRng>> {
+        Recrypt::new_with_rand(rand::thread_rng())
     }
 }
 
-impl Default for Api<Sha256, Ed25519, RandomBytes<rand::rngs::ThreadRng>> {
+impl Default for Recrypt<Sha256, Ed25519, RandomBytes<rand::rngs::ThreadRng>> {
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl<CR: rand::CryptoRng + rand::RngCore> Api<Sha256, Ed25519, RandomBytes<CR>> {
-    pub fn new_with_rand(r: CR) -> Api<Sha256, Ed25519, RandomBytes<CR>> {
+impl<CR: rand::CryptoRng + rand::RngCore> Recrypt<Sha256, Ed25519, RandomBytes<CR>> {
+    pub fn new_with_rand(r: CR) -> Recrypt<Sha256, Ed25519, RandomBytes<CR>> {
         let pairing = internal::pairing::Pairing::new();
         let curve_points = &*curve::FP_256_CURVE_POINTS;
         let schnorr_signing = internal::schnorr::SchnorrSign::<Monty256, Fr256, Sha256>::new_256();
-        Api {
+        Recrypt {
             random_bytes: RandomBytes::new(r),
             sha_256: Sha256,
             ed25519: Ed25519,
@@ -476,7 +477,7 @@ impl HashedValue {
             dest.copy_from_slice(bytes);
             Ok(HashedValue::new(dest)?)
         } else {
-            Err(ApiErr::InputWrongSize(
+            Err(RecryptErr::InputWrongSize(
                 "HashedValue",
                 HashedValue::ENCODED_SIZE_BYTES,
             ))
@@ -640,7 +641,7 @@ pub trait SchnorrOps {
     fn schnorr_sign<A: Hashable>(
         &mut self,
         priv_key: &PrivateKey,
-        pub_key: PublicKey,
+        pub_key: &PublicKey,
         message: &A,
     ) -> SchnorrSignature;
 
@@ -653,7 +654,7 @@ pub trait SchnorrOps {
     /// - `signature` - The signature that was generated from `schnorr_sign`.
     fn schnorr_verify<A: Hashable>(
         &self,
-        pub_key: PublicKey,
+        pub_key: &PublicKey,
         augmenting_priv_key: Option<&PrivateKey>,
         message: &A,
         signature: SchnorrSignature,
@@ -661,12 +662,12 @@ pub trait SchnorrOps {
 }
 
 impl<H: Sha256Hashing, S, CR: rand::RngCore + rand::CryptoRng> SchnorrOps
-    for Api<H, S, RandomBytes<CR>>
+    for Recrypt<H, S, RandomBytes<CR>>
 {
     fn schnorr_sign<A: Hashable>(
         &mut self,
         priv_key: &PrivateKey,
-        pub_key: PublicKey,
+        pub_key: &PublicKey,
         message: &A,
     ) -> SchnorrSignature {
         let k = Fr256::from_rand_no_bias(&mut self.random_bytes);
@@ -678,7 +679,7 @@ impl<H: Sha256Hashing, S, CR: rand::RngCore + rand::CryptoRng> SchnorrOps
 
     fn schnorr_verify<A: Hashable>(
         &self,
-        pub_key: PublicKey,
+        pub_key: &PublicKey,
         augmenting_priv_key: Option<&PrivateKey>,
         message: &A,
         signature: SchnorrSignature,
@@ -697,7 +698,7 @@ pub trait Ed25519Ops {
     fn generate_ed25519_key_pair(&mut self) -> SigningKeypair;
 }
 
-impl<H, S, CR: rand::RngCore + rand::CryptoRng> Ed25519Ops for Api<H, S, RandomBytes<CR>> {
+impl<H, S, CR: rand::RngCore + rand::CryptoRng> Ed25519Ops for Recrypt<H, S, RandomBytes<CR>> {
     ///Generate a signing key pair for use with the `Ed25519Signing` trait using the random number generator
     ///used to back the `RandomBytes` struct.
     fn generate_ed25519_key_pair(&mut self) -> SigningKeypair {
@@ -733,12 +734,12 @@ pub trait KeyGenOps {
     fn generate_transform_key(
         &mut self,
         from_private_key: &PrivateKey,
-        to_public_key: PublicKey,
+        to_public_key: &PublicKey,
         signing_keypair: &SigningKeypair,
     ) -> Result<TransformKey>;
 }
 
-impl<R: RandomBytesGen, H: Sha256Hashing, S: Ed25519Signing> KeyGenOps for Api<H, S, R> {
+impl<R: RandomBytesGen, H: Sha256Hashing, S: Ed25519Signing> KeyGenOps for Recrypt<H, S, R> {
     fn compute_public_key(&self, private_key: &PrivateKey) -> Result<PublicKey> {
         let pub_key_internal = internal::public_keygen(
             internal::PrivateKey::from(private_key),
@@ -761,7 +762,7 @@ impl<R: RandomBytesGen, H: Sha256Hashing, S: Ed25519Signing> KeyGenOps for Api<H
     fn generate_transform_key(
         &mut self,
         from_private_key: &PrivateKey,
-        to_public_key: PublicKey,
+        to_public_key: &PublicKey,
         signing_keypair: &SigningKeypair,
     ) -> Result<TransformKey> {
         let ephem_reencryption_private_key = self.random_private_key();
@@ -794,11 +795,12 @@ pub trait CryptoOps {
     fn gen_plaintext(&mut self) -> Plaintext;
 
     /// Convert our plaintext into a DecryptedSymmetricKey by hashing it.
+    /// Typically you either use `derive_private_key` or `derive_symmetric_key` but not both.
     fn derive_symmetric_key(&self, decrypted_value: &Plaintext) -> DerivedSymmetricKey;
 
-    ///Compute the stable hash of a value. This can be used to hash a Plaintext into a symmetric key or to generate a
-    ///PrivateKey from a Plaintext which you're encrypting to someone else.
-    fn hash_256<T: Hashable>(&self, to_hash: &T) -> [u8; 32];
+    ///Derive a private key for a plaintext by hashing it and modding it by the prime.
+    ///Typically you either use `derive_private_key` or `derive_symmetric_key` but not both.
+    fn derive_private_key(&self, plaintext: &Plaintext) -> PrivateKey;
 
     /// Encrypt the plaintext to the `to_public_key`.
     ///
@@ -808,11 +810,11 @@ pub trait CryptoOps {
     /// - `signing_keypair`       - signing keypair of the person (or device) who is encrypting this value
     ///
     /// # Return
-    /// EncryptedValue which can be decrypted by the matching private key of `to_public_key` or ApiErr.
+    /// EncryptedValue which can be decrypted by the matching private key of `to_public_key` or RecryptErr.
     fn encrypt(
         &mut self,
         plaintext: &Plaintext,
-        to_public_key: PublicKey,
+        to_public_key: &PublicKey,
         signing_keypair: &SigningKeypair,
     ) -> Result<EncryptedValue>;
 
@@ -843,24 +845,24 @@ pub trait CryptoOps {
     ) -> Result<EncryptedValue>;
 }
 
-impl<R: RandomBytesGen, H: Sha256Hashing, S: Ed25519Signing> CryptoOps for Api<H, S, R> {
+impl<R: RandomBytesGen, H: Sha256Hashing, S: Ed25519Signing> CryptoOps for Recrypt<H, S, R> {
     fn gen_plaintext(&mut self) -> Plaintext {
         let rand_fp12 = gen_random_fp12(&self.pairing, &mut self.random_bytes);
         Plaintext::from(rand_fp12)
     }
 
     fn derive_symmetric_key(&self, decrypted_value: &Plaintext) -> DerivedSymmetricKey {
-        DerivedSymmetricKey::new(self.hash_256(decrypted_value))
+        DerivedSymmetricKey::new(self.sha_256.hash(decrypted_value))
     }
 
-    fn hash_256<T: Hashable>(&self, to_hash: &T) -> [u8; 32] {
-        self.sha_256.hash(to_hash)
+    fn derive_private_key(&self, plaintext: &Plaintext) -> PrivateKey {
+        PrivateKey::new(self.sha_256.hash(plaintext))
     }
 
     fn encrypt(
         &mut self,
         plaintext: &Plaintext,
-        to_public_key: PublicKey,
+        to_public_key: &PublicKey,
         signing_keypair: &SigningKeypair,
     ) -> Result<EncryptedValue> {
         //generate a ephemeral private key
@@ -994,7 +996,7 @@ impl PublicKey {
 
             Ok(PublicKey::new((x_dest, y_dest))?)
         } else {
-            Err(ApiErr::InputWrongSize(
+            Err(RecryptErr::InputWrongSize(
                 "PublicKey",
                 PublicKey::ENCODED_SIZE_BYTES,
             ))
@@ -1020,7 +1022,7 @@ impl PartialEq for PublicKey {
     }
 }
 
-#[derive(Default, Debug)]
+#[derive(Default, Debug, Clone)]
 #[cfg_attr(test, derive(PartialEq))]
 // If you are looking for PartialEq for PrivateKey, see PartialEq for Revealed<PrivateKey>
 pub struct PrivateKey {
@@ -1029,7 +1031,7 @@ pub struct PrivateKey {
 }
 
 impl PrivateKey {
-    const ENCODED_SIZE_BYTES: usize = Monty256::ENCODED_SIZE_BYTES;
+    pub const ENCODED_SIZE_BYTES: usize = Monty256::ENCODED_SIZE_BYTES;
 
     pub fn bytes(&self) -> &[u8; PrivateKey::ENCODED_SIZE_BYTES] {
         &self.bytes
@@ -1131,9 +1133,9 @@ pub(crate) mod test {
     fn api_with<R: RandomBytesGen + Default, S: Ed25519Signing>(
         random_bytes: Option<R>,
         ed25519: S,
-    ) -> Api<Sha256, S, R> {
-        let api = Api::new();
-        Api::<Sha256, S, R> {
+    ) -> Recrypt<Sha256, S, R> {
+        let api = Recrypt::new();
+        Recrypt::<Sha256, S, R> {
             random_bytes: random_bytes.unwrap_or_default(),
             schnorr_signing: internal::schnorr::SchnorrSign::new_256(),
             sha_256: api.sha_256,
@@ -1145,28 +1147,28 @@ pub(crate) mod test {
 
     #[test]
     fn schnorr_signing_roundtrip_augmented() {
-        let mut api = Api::new();
+        let mut api = Recrypt::new();
         let (private_key, pub_key) = api.generate_key_pair().unwrap();
         let (aug_private_key, aug_pub_key) = api.generate_key_pair().unwrap();
         let message = vec![1u8, 2u8];
         let augmented_pub = pub_key.augment(&aug_pub_key).unwrap();
-        let sig = api.schnorr_sign(&private_key, augmented_pub, &message);
-        let result = api.schnorr_verify(augmented_pub, Some(&aug_private_key), &message, sig);
+        let sig = api.schnorr_sign(&private_key, &augmented_pub, &message);
+        let result = api.schnorr_verify(&augmented_pub, Some(&aug_private_key), &message, sig);
         assert!(result);
     }
     #[test]
     fn schnorr_signing_roundtrip_unaugmented() {
-        let mut api = Api::new();
+        let mut api = Recrypt::new();
         let (private_key, pub_key) = api.generate_key_pair().unwrap();
         let message = vec![1u8, 2u8, 3u8, 4u8];
-        let sig = api.schnorr_sign(&private_key, pub_key, &message);
-        let result = api.schnorr_verify(pub_key, None, &message, sig);
+        let sig = api.schnorr_sign(&private_key, &pub_key, &message);
+        let result = api.schnorr_verify(&pub_key, None, &message, sig);
         assert!(result);
     }
 
     #[test]
     fn public_key_roundtrip_with_internal() {
-        let (_, pub_key_api) = Api::new().generate_key_pair().unwrap();
+        let (_, pub_key_api) = Recrypt::new().generate_key_pair().unwrap();
 
         let internal_pk = pub_key_api._internal_key;
         let roundtrip = PublicKey::try_from(&internal_pk).unwrap();
@@ -1176,7 +1178,7 @@ pub(crate) mod test {
 
     #[test]
     fn private_key_roundtrip_with_internal() {
-        let (priv_key_api, _) = Api::new().generate_key_pair().unwrap();
+        let (priv_key_api, _) = Recrypt::new().generate_key_pair().unwrap();
 
         let internal_pk = internal::PrivateKey::<Monty256>::from(&priv_key_api);
         let roundtrip = PrivateKey::from(internal_pk);
@@ -1187,7 +1189,7 @@ pub(crate) mod test {
 
     #[test]
     fn gen_plaintext_len() {
-        let api = &mut Api::new();
+        let api = &mut Recrypt::new();
 
         let result = api.gen_plaintext();
         assert_eq!(Fp12Elem::<Monty256>::ENCODED_SIZE_BYTES, result.bytes.len());
@@ -1195,7 +1197,7 @@ pub(crate) mod test {
 
     #[test]
     fn test_compute_public_key() {
-        let api = &mut Api::new();
+        let api = &mut Recrypt::new();
         //37777967648492203239675772600961898148040325589588086812374811831221462604944
         let parsed_priv_key =
             fp256_unsafe_from("5385926b9f6135086d1912901e5a433ffcebc19a30fadbd0ee8cee26ba719c90")
@@ -1258,10 +1260,10 @@ pub(crate) mod test {
     }
 
     fn good_transform_key() -> TransformKey {
-        let mut api = Api::new();
+        let mut api = Recrypt::new();
         let signing_key = ed25519::test::good_signing_keypair();
         let (master_priv, master_pub) = api.generate_key_pair().unwrap();
-        api.generate_transform_key(&master_priv, master_pub, &signing_key)
+        api.generate_transform_key(&master_priv, &master_pub, &signing_key)
             .unwrap()
     }
 
@@ -1287,7 +1289,7 @@ pub(crate) mod test {
 
     #[test]
     fn roundtrip_transform_block() {
-        let mut api = Api::new();
+        let mut api = Recrypt::new();
         let pub_key1 = api.generate_key_pair().unwrap().1;
         let pub_key2 = api.generate_key_pair().unwrap().1;
         let ee1 = EncryptedTempKey::new(api.gen_plaintext().bytes);
@@ -1313,7 +1315,7 @@ pub(crate) mod test {
         let priv_key = PrivateKey::new_from_slice(&hex::decode(
             "3f79bb7b435b05321651daefd374cdc681dc06faa65e374e38337b88ca046dea",
         )?)?;
-        let api = Api::new();
+        let api = Recrypt::new();
         let pt = api.decrypt(encrypted, &priv_key)?;
         assert_eq!(pt, expected_pt);
         Ok(())
@@ -1322,12 +1324,12 @@ pub(crate) mod test {
     #[test]
     fn encrypt_decrypt_roundtrip() -> Result<()> {
         use rand::SeedableRng;
-        let mut api = Api::new_with_rand(rand_chacha::ChaChaRng::from_seed([0u8; 32]));
+        let mut api = Recrypt::new_with_rand(rand_chacha::ChaChaRng::from_seed([0u8; 32]));
         let pt = api.gen_plaintext();
         let (priv_key, pub_key) = api.generate_key_pair().unwrap();
         let priv_signing_key = api.generate_ed25519_key_pair();
 
-        let encrypted_val = api.encrypt(&pt, pub_key, &priv_signing_key).unwrap();
+        let encrypted_val = api.encrypt(&pt, &pub_key, &priv_signing_key).unwrap();
 
         let decrypted_val = api.decrypt(encrypted_val, &priv_key).unwrap();
 
@@ -1345,8 +1347,27 @@ pub(crate) mod test {
         let mut dest: [u8; 32] = [0u8; 32];
         dest.copy_from_slice(src);
         let expected_result = DerivedSymmetricKey::new(dest);
-        let result = Api::new().derive_symmetric_key(&pt);
+        let result = Recrypt::new().derive_symmetric_key(&pt);
         assert_eq!(Revealed(expected_result), Revealed(result))
+    }
+
+    #[test]
+    ///This test is meant to show that if the top byte is too large the private key and derived symmetric key might actually be different.
+    fn derive_known_symmetric_key_not_same_bytes_as_private_key() {
+        let api = Recrypt::default();
+        let bytes = hex::decode("34f36d6fb086b38435823c96f016fc8e41c7ab39c1abb02a773333b88f8d1f1409289fccaa485629e15d3273768e2a039368c68dc5873353b5c18a2d0eb02adf04519ded7fc4bd07c2b405b8f9075b96be28915f793f3e90b3e5488f20c666ff00839462c603d7f1f9d5c86556a0590bc2a76fb5d2d6dc2afa53fb5470af3a5521dd82ee76290502a84a0bc5e7e37b183332dc761fb808b8e7ba138cbee30a802f9257b5f2117452025a1e92e45b9624ad29f46db639f223e7c067e1fdb3c93d4f55165a15ec90451272325f19678f2d3e6230736916ec562fbda94f920d5149506b7efe1211ac62e826f1b8d2f8c41f10c1cf4d53a7222d5124b536c3707b0b86198131f9f4ef2cfdf7ff9d13bc6b6e21f8e0a337a0acda48055d10143381760e783473a14153b371b1147c18852acb4af0a3d4d9dd7e738b04e7cd0c0a6b5a1b826f3aa4817cfab2ccb73ab03258e42b7baa54cde8a903de4d3a6b8c7742e92b9976fdf64c496dab1d143f4d65bc86d9f8f6e3ee38e97da3faa8bbdf461688").unwrap();
+        let pt = Plaintext::from(Fp12Elem::decode(bytes).unwrap());
+        //This is a manually computed symmetric key based on the above plaintext (which was chosen because this value is greater than P)
+        let src = &hex::decode("cd1b366b2575f2a69390c51b3b1e0c3e2eace761e0a4cee2a1895175071f6700")
+            .unwrap()[..];
+        let mut dest: [u8; 32] = [0u8; 32];
+        dest.copy_from_slice(src);
+        let expected_result = DerivedSymmetricKey::new(dest);
+        let result = api.derive_symmetric_key(&pt);
+        assert_eq!(Revealed(expected_result), Revealed(result));
+        //This hashes, but also mods the value so it's not the same.
+        let private_key_result = api.derive_private_key(&pt);
+        assert_ne!(private_key_result.bytes(), result.bytes());
     }
 
     use std::default::Default;
@@ -1357,9 +1378,9 @@ pub(crate) mod test {
 
         let plaintext = api.gen_plaintext();
         let (master_priv, master_pub) = api.generate_key_pair().unwrap();
-        let enc_value = api.encrypt(&plaintext, master_pub, &signing_key).unwrap();
+        let enc_value = api.encrypt(&plaintext, &master_pub, &signing_key).unwrap();
         let master_to_master_transform_key = api
-            .generate_transform_key(&master_priv, master_pub, &signing_key)
+            .generate_transform_key(&master_priv, &master_pub, &signing_key)
             .unwrap();
         let transformed_value = api
             .transform(enc_value, master_to_master_transform_key, &signing_key)
@@ -1376,9 +1397,9 @@ pub(crate) mod test {
         let pt = api.gen_plaintext();
         let (master_private_key, master_public_key) = api.generate_key_pair().unwrap();
         let (device_private_key, device_public_key) = api.generate_key_pair().unwrap();
-        let encrypted_msg = api.encrypt(&pt, master_public_key, &signing_key).unwrap();
+        let encrypted_msg = api.encrypt(&pt, &master_public_key, &signing_key).unwrap();
         let master_to_device_transform_key = api
-            .generate_transform_key(&master_private_key, device_public_key, &signing_key)
+            .generate_transform_key(&master_private_key, &device_public_key, &signing_key)
             .unwrap();
 
         let transformed_msg = api
@@ -1391,16 +1412,16 @@ pub(crate) mod test {
 
     #[test]
     fn encrypt_decrypt_roundtrip_augmented_keys() {
-        let mut api = Api::new();
+        let mut api = Recrypt::new();
         let signing_key = api.generate_ed25519_key_pair();
         let pt = api.gen_plaintext();
         let (master_private_key, client_generated_pub) = api.generate_key_pair().unwrap();
         let (device_private_key, device_public_key) = api.generate_key_pair().unwrap();
         let (server_private, server_public) = api.generate_key_pair().unwrap();
         let master_public_key = client_generated_pub.augment(&server_public).unwrap();
-        let encrypted_msg = api.encrypt(&pt, master_public_key, &signing_key).unwrap();
+        let encrypted_msg = api.encrypt(&pt, &master_public_key, &signing_key).unwrap();
         let master_to_device_transform_key = api
-            .generate_transform_key(&master_private_key, device_public_key, &signing_key)
+            .generate_transform_key(&master_private_key, &device_public_key, &signing_key)
             .unwrap();
         let augmented_transform_key = master_to_device_transform_key
             .augment(&server_private)
@@ -1423,20 +1444,20 @@ pub(crate) mod test {
         let (user_master_private_key, user_master_public_key) = api.generate_key_pair().unwrap();
         let (device_private_key, device_public_key) = api.generate_key_pair().unwrap();
         let encrypted_msg = api
-            .encrypt(&pt, group_master_public_key, &signing_key)
+            .encrypt(&pt, &group_master_public_key, &signing_key)
             .unwrap();
 
         // now create two transform keys. Group -> User -> Device (arrows are the transform keys)
         let group_to_user_transform_key = api
             .generate_transform_key(
                 &group_master_private_key,
-                user_master_public_key,
+                &user_master_public_key,
                 &signing_key,
             )
             .unwrap();
 
         let user_to_device_transform_key = api
-            .generate_transform_key(&user_master_private_key, device_public_key, &signing_key)
+            .generate_transform_key(&user_master_private_key, &device_public_key, &signing_key)
             .unwrap();
 
         let transformed_to_user = api
@@ -1459,7 +1480,7 @@ pub(crate) mod test {
     #[test]
     fn generate_ed25519_key_pair() {
         use rand::SeedableRng;
-        let mut api = Api::new_with_rand(rand_chacha::ChaChaRng::from_seed([0u8; 32]));
+        let mut api = Recrypt::new_with_rand(rand_chacha::ChaChaRng::from_seed([0u8; 32]));
         let signing_keypair = Revealed(api.generate_ed25519_key_pair());
         let expected_signing_keypair = Revealed(SigningKeypair::new_unchecked([
             118, 184, 224, 173, 160, 241, 61, 144, 64, 93, 106, 229, 83, 134, 189, 40, 189, 210,
@@ -1490,7 +1511,7 @@ pub(crate) mod test {
         assert_eq!(auth_hash_from_fixed, auth_hash_from_slice.unwrap());
 
         assert_eq!(
-            ApiErr::InputWrongSize("AuthHash", 32),
+            RecryptErr::InputWrongSize("AuthHash", 32),
             AuthHash::new_from_slice(&input[..30]).unwrap_err()
         )
     }
@@ -1505,13 +1526,13 @@ pub(crate) mod test {
         assert_eq!(hv_from_fixed.unwrap(), hv_from_slice.unwrap());
 
         assert_eq!(
-            ApiErr::InputWrongSize("HashedValue", 128),
+            RecryptErr::InputWrongSize("HashedValue", 128),
             HashedValue::new_from_slice(&input[..30]).unwrap_err()
         )
     }
     #[test]
     fn publickey_new_from_slice() {
-        let mut api = Api::new();
+        let mut api = Recrypt::new();
         let (_, pk1) = api.generate_key_pair().unwrap();
         let input: ([u8; 32], [u8; 32]) = (pk1.x, pk1.y);
         let slice: (&[u8], &[u8]) = (&input.0, &input.1);
@@ -1521,15 +1542,31 @@ pub(crate) mod test {
         assert_eq!(pk_from_fixed.unwrap(), pk_from_slice.unwrap());
 
         assert_eq!(
-            ApiErr::InputWrongSize("PublicKey", 64),
+            RecryptErr::InputWrongSize("PublicKey", 64),
             PublicKey::new_from_slice((&input.0[..30], &input.1[..32])).unwrap_err()
+        )
+    }
+
+    #[test]
+    fn private_key_new_from_slice() {
+        let mut rand_bytes = DummyRandomBytes;
+        let input: [u8; 32] = rand_bytes.random_bytes_32();
+        let slice: &[u8] = &input;
+        let from_fixed = PrivateKey::new(input);
+        let from_slice = PrivateKey::new_from_slice(slice);
+
+        assert_eq!(from_fixed, from_slice.unwrap());
+
+        assert_eq!(
+            RecryptErr::InputWrongSize("PrivateKey", 32),
+            PrivateKey::new_from_slice(&input[..30]).unwrap_err()
         )
     }
 
     // note that this doesn't show that Drop is working properly, just that clear does
     #[test]
     fn private_key_clear() {
-        let (mut priv_key, _) = Api::new().generate_key_pair().unwrap();
+        let (mut priv_key, _) = Recrypt::new().generate_key_pair().unwrap();
         priv_key.clear();
         assert_eq!(priv_key.bytes(), &[0u8; 32]);
         assert_eq!(priv_key._internal_key, Default::default())