+++ src/api_480.rs
@@ -1,4 +1,5 @@
-use crate::api_common::{ApiErr, Result};
+use crate::api_common::RecryptErr;
+use crate::api_common::Result;
 use crate::internal;
 use crate::internal::bytedecoder::{BytesDecoder, DecodeErr};
 use crate::internal::curve;
@@ -27,7 +28,7 @@ use std::fmt;
 /// Recrypt public API - 480-bit
 /// If you are looking better performance, you might consider the 256-bit API in `api.rs`
 #[derive(Debug)]
-pub struct Api480<H, S, R> {
+pub struct Recrypt480<H, S, R> {
     random_bytes: R,
     sha_256: H,
     ed25519: S,
@@ -36,24 +37,24 @@ pub struct Api480<H, S, R> {
     schnorr_signing: SchnorrSign<Monty480, Fr480, H>,
 }
 
-impl Api480<Sha256, Ed25519, RandomBytes<rand::rngs::ThreadRng>> {
-    pub fn new() -> Api480<Sha256, Ed25519, RandomBytes<rand::rngs::ThreadRng>> {
-        Api480::new_with_rand(rand::thread_rng())
+impl Recrypt480<Sha256, Ed25519, RandomBytes<rand::rngs::ThreadRng>> {
+    pub fn new() -> Recrypt480<Sha256, Ed25519, RandomBytes<rand::rngs::ThreadRng>> {
+        Recrypt480::new_with_rand(rand::thread_rng())
     }
 }
 
-impl Default for Api480<Sha256, Ed25519, RandomBytes<rand::rngs::ThreadRng>> {
+impl Default for Recrypt480<Sha256, Ed25519, RandomBytes<rand::rngs::ThreadRng>> {
     fn default() -> Self {
         Self::new()
     }
 }
 
-impl<CR: rand::CryptoRng + rand::RngCore> Api480<Sha256, Ed25519, RandomBytes<CR>> {
-    pub fn new_with_rand(r: CR) -> Api480<Sha256, Ed25519, RandomBytes<CR>> {
+impl<CR: rand::CryptoRng + rand::RngCore> Recrypt480<Sha256, Ed25519, RandomBytes<CR>> {
+    pub fn new_with_rand(r: CR) -> Recrypt480<Sha256, Ed25519, RandomBytes<CR>> {
         let pairing = pairing::Pairing::new();
         let curve_points = &*curve::FP_480_CURVE_POINTS;
         let schnorr_signing = internal::schnorr::SchnorrSign::<Monty480, Fr480, Sha256>::new_480();
-        Api480 {
+        Recrypt480 {
             random_bytes: RandomBytes::new(r),
             sha_256: Sha256,
             ed25519: Ed25519,
@@ -477,7 +478,7 @@ impl HashedValue {
             dest.copy_from_slice(bytes);
             Ok(HashedValue::new(dest)?)
         } else {
-            Err(ApiErr::InputWrongSize(
+            Err(RecryptErr::InputWrongSize(
                 "HashedValue",
                 HashedValue::ENCODED_SIZE_BYTES,
             ))
@@ -641,7 +642,7 @@ pub trait SchnorrOps {
     fn schnorr_sign<A: Hashable>(
         &mut self,
         priv_key: &PrivateKey,
-        pub_key: PublicKey,
+        pub_key: &PublicKey,
         message: &A,
     ) -> SchnorrSignature;
 
@@ -654,7 +655,7 @@ pub trait SchnorrOps {
     /// - `signature` - The signature that was generated from `schnorr_sign`.
     fn schnorr_verify<A: Hashable>(
         &self,
-        pub_key: PublicKey,
+        pub_key: &PublicKey,
         augmenting_priv_key: Option<&PrivateKey>,
         message: &A,
         signature: SchnorrSignature,
@@ -662,12 +663,12 @@ pub trait SchnorrOps {
 }
 
 impl<H: Sha256Hashing, S, CR: rand::RngCore + rand::CryptoRng> SchnorrOps
-    for Api480<H, S, RandomBytes<CR>>
+    for Recrypt480<H, S, RandomBytes<CR>>
 {
     fn schnorr_sign<A: Hashable>(
         &mut self,
         priv_key: &PrivateKey,
-        pub_key: PublicKey,
+        pub_key: &PublicKey,
         message: &A,
     ) -> SchnorrSignature {
         let k = Fr480::from_rand_no_bias(&mut self.random_bytes);
@@ -679,7 +680,7 @@ impl<H: Sha256Hashing, S, CR: rand::RngCore + rand::CryptoRng> SchnorrOps
 
     fn schnorr_verify<A: Hashable>(
         &self,
-        pub_key: PublicKey,
+        pub_key: &PublicKey,
         augmenting_priv_key: Option<&PrivateKey>,
         message: &A,
         signature: SchnorrSignature,
@@ -698,7 +699,7 @@ pub trait Ed25519Ops {
     fn generate_ed25519_key_pair(&mut self) -> SigningKeypair;
 }
 
-impl<H, S, CR: rand::RngCore + rand::CryptoRng> Ed25519Ops for Api480<H, S, RandomBytes<CR>> {
+impl<H, S, CR: rand::RngCore + rand::CryptoRng> Ed25519Ops for Recrypt480<H, S, RandomBytes<CR>> {
     ///Generate a signing key pair for use with the `Ed25519Signing` trait using the random number generator
     ///used to back the `RandomBytes` struct.
     fn generate_ed25519_key_pair(&mut self) -> SigningKeypair {
@@ -734,12 +735,12 @@ pub trait KeyGenOps {
     fn generate_transform_key(
         &mut self,
         from_private_key: &PrivateKey,
-        to_public_key: PublicKey,
+        to_public_key: &PublicKey,
         signing_keypair: &SigningKeypair,
     ) -> Result<TransformKey>;
 }
 
-impl<R: RandomBytesGen, H: Sha256Hashing, S: Ed25519Signing> KeyGenOps for Api480<H, S, R> {
+impl<R: RandomBytesGen, H: Sha256Hashing, S: Ed25519Signing> KeyGenOps for Recrypt480<H, S, R> {
     fn compute_public_key(&self, private_key: &PrivateKey) -> Result<PublicKey> {
         let pub_key_internal = internal::public_keygen(
             internal::PrivateKey::from(private_key),
@@ -762,7 +763,7 @@ impl<R: RandomBytesGen, H: Sha256Hashing, S: Ed25519Signing> KeyGenOps for Api48
     fn generate_transform_key(
         &mut self,
         from_private_key: &PrivateKey,
-        to_public_key: PublicKey,
+        to_public_key: &PublicKey,
         signing_keypair: &SigningKeypair,
     ) -> Result<TransformKey> {
         let ephem_reencryption_private_key = self.random_private_key();
@@ -797,9 +798,9 @@ pub trait CryptoOps {
     /// Convert our plaintext into a DecryptedSymmetricKey by hashing it.
     fn derive_symmetric_key(&self, decrypted_value: &Plaintext) -> DerivedSymmetricKey;
 
-    ///Compute the stable hash of a value. This can be used to hash a Plaintext into a symmetric key or to generate a
-    ///PrivateKey from a Plaintext which you're encrypting to someone else.
-    fn hash_256<T: Hashable>(&self, to_hash: &T) -> [u8; 32];
+    ///Derive a private key for a plaintext by hashing it twice (with known leading bytes) and modding it by the prime.
+    ///Typically you either use `derive_private_key` or `derive_symmetric_key` but not both.
+    fn derive_private_key(&self, plaintext: &Plaintext) -> PrivateKey;
 
     /// Encrypt the plaintext to the `to_public_key`.
     ///
@@ -809,11 +810,11 @@ pub trait CryptoOps {
     /// - `signing_keypair`       - signing keypair of the person (or device) who is encrypting this value
     ///
     /// # Return
-    /// EncryptedValue which can be decrypted by the matching private key of `to_public_key` or ApiErr.
+    /// EncryptedValue which can be decrypted by the matching private key of `to_public_key` or RecryptErr.
     fn encrypt(
         &mut self,
         plaintext: &Plaintext,
-        to_public_key: PublicKey,
+        to_public_key: &PublicKey,
         signing_keypair: &SigningKeypair,
     ) -> Result<EncryptedValue>;
 
@@ -844,24 +845,30 @@ pub trait CryptoOps {
     ) -> Result<EncryptedValue>;
 }
 
-impl<R: RandomBytesGen, H: Sha256Hashing, S: Ed25519Signing> CryptoOps for Api480<H, S, R> {
+impl<R: RandomBytesGen, H: Sha256Hashing, S: Ed25519Signing> CryptoOps for Recrypt480<H, S, R> {
     fn gen_plaintext(&mut self) -> Plaintext {
         let rand_fp12 = gen_random_fp12(&mut self.random_bytes);
         Plaintext::from(rand_fp12)
     }
 
     fn derive_symmetric_key(&self, decrypted_value: &Plaintext) -> DerivedSymmetricKey {
-        DerivedSymmetricKey::new(self.hash_256(decrypted_value))
+        DerivedSymmetricKey::new(self.sha_256.hash(decrypted_value))
     }
 
-    fn hash_256<T: Hashable>(&self, to_hash: &T) -> [u8; 32] {
-        self.sha_256.hash(to_hash)
+    fn derive_private_key(&self, plaintext: &Plaintext) -> PrivateKey {
+        let mut result_bytes = [0u8; 60];
+
+        //we're defining the mapping of plaintext to private key as 0 and plaintext hashed combined with 1 and plaintext hashed.
+        //We then chop it to 60 and let the mod happen in the private key constructor.
+        result_bytes[0..28].copy_from_slice(&self.sha_256.hash(&(&0u8, plaintext))[4..]);
+        result_bytes[28..60].copy_from_slice(&self.sha_256.hash(&(&1u8, plaintext))[..]);
+        PrivateKey::new(result_bytes)
     }
 
     fn encrypt(
         &mut self,
         plaintext: &Plaintext,
-        to_public_key: PublicKey,
+        to_public_key: &PublicKey,
         signing_keypair: &SigningKeypair,
     ) -> Result<EncryptedValue> {
         //generate a ephemeral private key
@@ -1020,7 +1027,7 @@ impl PublicKey {
 
             Ok(PublicKey::new((x_dest, y_dest))?)
         } else {
-            Err(ApiErr::InputWrongSize(
+            Err(RecryptErr::InputWrongSize(
                 "PublicKey",
                 PublicKey::ENCODED_SIZE_BYTES,
             ))
@@ -1051,7 +1058,7 @@ impl PartialEq for PublicKey {
     }
 }
 
-#[derive(Default, Debug)]
+#[derive(Default, Debug, Clone)]
 #[cfg_attr(test, derive(PartialEq))]
 // If you are looking for PartialEq for PrivateKey, see PartialEq for Revealed<PrivateKey>
 pub struct PrivateKey {
@@ -1060,7 +1067,7 @@ pub struct PrivateKey {
 }
 
 impl PrivateKey {
-    const ENCODED_SIZE_BYTES: usize = Monty480::ENCODED_SIZE_BYTES;
+    pub const ENCODED_SIZE_BYTES: usize = Monty480::ENCODED_SIZE_BYTES;
 
     pub fn bytes(&self) -> &[u8; PrivateKey::ENCODED_SIZE_BYTES] {
         &self.bytes.0
@@ -1167,9 +1174,9 @@ pub(crate) mod test {
     fn api_with<R: RandomBytesGen + Default, S: Ed25519Signing>(
         random_bytes: Option<R>,
         ed25519: S,
-    ) -> Api480<Sha256, S, R> {
-        let api = Api480::new();
-        Api480 {
+    ) -> Recrypt480<Sha256, S, R> {
+        let api = Recrypt480::new();
+        Recrypt480 {
             random_bytes: random_bytes.unwrap_or_default(),
             sha_256: api.sha_256,
             ed25519,
@@ -1181,28 +1188,28 @@ pub(crate) mod test {
 
     #[test]
     fn schnorr_signing_roundtrip_augmented() {
-        let mut api = Api480::new();
+        let mut api = Recrypt480::new();
         let (private_key, pub_key) = api.generate_key_pair().unwrap();
         let (aug_private_key, aug_pub_key) = api.generate_key_pair().unwrap();
         let message = vec![1u8, 2u8];
         let augmented_pub = pub_key.augment(&aug_pub_key).unwrap();
-        let sig = api.schnorr_sign(&private_key, augmented_pub, &message);
-        let result = api.schnorr_verify(augmented_pub, Some(&aug_private_key), &message, sig);
+        let sig = api.schnorr_sign(&private_key, &augmented_pub, &message);
+        let result = api.schnorr_verify(&augmented_pub, Some(&aug_private_key), &message, sig);
         assert!(result);
     }
     #[test]
     fn schnorr_signing_roundtrip_unaugmented() {
-        let mut api = Api480::new();
+        let mut api = Recrypt480::new();
         let (private_key, pub_key) = api.generate_key_pair().unwrap();
         let message = vec![1u8, 2u8, 3u8, 4u8];
-        let sig = api.schnorr_sign(&private_key, pub_key, &message);
-        let result = api.schnorr_verify(pub_key, None, &message, sig);
+        let sig = api.schnorr_sign(&private_key, &pub_key, &message);
+        let result = api.schnorr_verify(&pub_key, None, &message, sig);
         assert!(result);
     }
 
     #[test]
     fn public_key_roundtrip_with_internal() {
-        let (_, pub_key_api) = Api480::new().generate_key_pair().unwrap();
+        let (_, pub_key_api) = Recrypt480::new().generate_key_pair().unwrap();
 
         let internal_pk = pub_key_api._internal_key;
         let roundtrip = PublicKey::try_from(&internal_pk).unwrap();
@@ -1212,7 +1219,7 @@ pub(crate) mod test {
 
     #[test]
     fn private_key_roundtrip_with_internal() {
-        let (priv_key_api, _) = Api480::new().generate_key_pair().unwrap();
+        let (priv_key_api, _) = Recrypt480::new().generate_key_pair().unwrap();
 
         let internal_pk = internal::PrivateKey::<Monty480>::from(&priv_key_api);
         let roundtrip = PrivateKey::from(internal_pk);
@@ -1223,7 +1230,7 @@ pub(crate) mod test {
 
     #[test]
     fn gen_plaintext_len() {
-        let api = &mut Api480::new();
+        let api = &mut Recrypt480::new();
 
         let result = api.gen_plaintext();
         assert_eq!(Fp12Elem::<Monty480>::ENCODED_SIZE_BYTES, result.bytes.len());
@@ -1261,10 +1268,10 @@ pub(crate) mod test {
     }
 
     fn good_transform_key() -> TransformKey {
-        let mut api = Api480::new();
+        let mut api = Recrypt480::new();
         let signing_key = ed25519::test::good_signing_keypair();
         let (master_priv, master_pub) = api.generate_key_pair().unwrap();
-        api.generate_transform_key(&master_priv, master_pub, &signing_key)
+        api.generate_transform_key(&master_priv, &master_pub, &signing_key)
             .unwrap()
     }
 
@@ -1290,7 +1297,7 @@ pub(crate) mod test {
 
     #[test]
     fn roundtrip_transform_block() {
-        let mut api = Api480::new();
+        let mut api = Recrypt480::new();
         let pub_key1 = api.generate_key_pair().unwrap().1;
         let pub_key2 = api.generate_key_pair().unwrap().1;
         let ee1 = EncryptedTempKey::new(api.gen_plaintext().bytes);
@@ -1306,12 +1313,12 @@ pub(crate) mod test {
     #[test]
     fn encrypt_decrypt_roundtrip() -> Result<()> {
         use rand::SeedableRng;
-        let mut api = Api480::new_with_rand(rand_chacha::ChaChaRng::from_seed([0u8; 32]));
+        let mut api = Recrypt480::new_with_rand(rand_chacha::ChaChaRng::from_seed([0u8; 32]));
         let pt = api.gen_plaintext();
         let (priv_key, pub_key) = api.generate_key_pair().unwrap();
         let priv_signing_key = api.generate_ed25519_key_pair();
 
-        let encrypted_val = api.encrypt(&pt, pub_key, &priv_signing_key).unwrap();
+        let encrypted_val = api.encrypt(&pt, &pub_key, &priv_signing_key).unwrap();
 
         let decrypted_val = api.decrypt(encrypted_val, &priv_key).unwrap();
 
@@ -1328,9 +1335,9 @@ pub(crate) mod test {
 
         let plaintext = api.gen_plaintext();
         let (master_priv, master_pub) = api.generate_key_pair().unwrap();
-        let enc_value = api.encrypt(&plaintext, master_pub, &signing_key).unwrap();
+        let enc_value = api.encrypt(&plaintext, &master_pub, &signing_key).unwrap();
         let master_to_master_transform_key = api
-            .generate_transform_key(&master_priv, master_pub, &signing_key)
+            .generate_transform_key(&master_priv, &master_pub, &signing_key)
             .unwrap();
         let transformed_value = api
             .transform(enc_value, master_to_master_transform_key, &signing_key)
@@ -1347,9 +1354,9 @@ pub(crate) mod test {
         let pt = api.gen_plaintext();
         let (master_private_key, master_public_key) = api.generate_key_pair().unwrap();
         let (device_private_key, device_public_key) = api.generate_key_pair().unwrap();
-        let encrypted_msg = api.encrypt(&pt, master_public_key, &signing_key).unwrap();
+        let encrypted_msg = api.encrypt(&pt, &master_public_key, &signing_key).unwrap();
         let master_to_device_transform_key = api
-            .generate_transform_key(&master_private_key, device_public_key, &signing_key)
+            .generate_transform_key(&master_private_key, &device_public_key, &signing_key)
             .unwrap();
 
         let transformed_msg = api
@@ -1362,16 +1369,16 @@ pub(crate) mod test {
 
     #[test]
     fn encrypt_decrypt_roundtrip_augmented_keys() {
-        let mut api = Api480::new();
+        let mut api = Recrypt480::new();
         let signing_key = api.generate_ed25519_key_pair();
         let pt = api.gen_plaintext();
         let (master_private_key, client_generated_pub) = api.generate_key_pair().unwrap();
         let (device_private_key, device_public_key) = api.generate_key_pair().unwrap();
         let (server_private, server_public) = api.generate_key_pair().unwrap();
         let master_public_key = client_generated_pub.augment(&server_public).unwrap();
-        let encrypted_msg = api.encrypt(&pt, master_public_key, &signing_key).unwrap();
+        let encrypted_msg = api.encrypt(&pt, &master_public_key, &signing_key).unwrap();
         let master_to_device_transform_key = api
-            .generate_transform_key(&master_private_key, device_public_key, &signing_key)
+            .generate_transform_key(&master_private_key, &device_public_key, &signing_key)
             .unwrap();
         let augmented_transform_key = master_to_device_transform_key
             .augment(&server_private)
@@ -1394,20 +1401,20 @@ pub(crate) mod test {
         let (user_master_private_key, user_master_public_key) = api.generate_key_pair().unwrap();
         let (device_private_key, device_public_key) = api.generate_key_pair().unwrap();
         let encrypted_msg = api
-            .encrypt(&pt, group_master_public_key, &signing_key)
+            .encrypt(&pt, &group_master_public_key, &signing_key)
             .unwrap();
 
         // now create two transform keys. Group -> User -> Device (arrows are the transform keys)
         let group_to_user_transform_key = api
             .generate_transform_key(
                 &group_master_private_key,
-                user_master_public_key,
+                &user_master_public_key,
                 &signing_key,
             )
             .unwrap();
 
         let user_to_device_transform_key = api
-            .generate_transform_key(&user_master_private_key, device_public_key, &signing_key)
+            .generate_transform_key(&user_master_private_key, &device_public_key, &signing_key)
             .unwrap();
 
         let transformed_to_user = api
@@ -1438,7 +1445,7 @@ pub(crate) mod test {
         assert_eq!(auth_hash_from_fixed, auth_hash_from_slice.unwrap());
 
         assert_eq!(
-            ApiErr::InputWrongSize("AuthHash", 32),
+            RecryptErr::InputWrongSize("AuthHash", 32),
             AuthHash::new_from_slice(&input[..30]).unwrap_err()
         )
     }
@@ -1453,13 +1460,27 @@ pub(crate) mod test {
         assert_eq!(hv_from_fixed.unwrap(), hv_from_slice.unwrap());
 
         assert_eq!(
-            ApiErr::InputWrongSize("HashedValue", 240),
+            RecryptErr::InputWrongSize("HashedValue", 240),
             HashedValue::new_from_slice(&input[..30]).unwrap_err()
         )
     }
+
+    #[test]
+    fn derive_private_key_is_known_value() {
+        let api = api_with(Some(RandomBytes::default()), DummyEd25519);
+        let bytes = hex::decode("a5e480feba36757014d53c9efd823cd69ae0f8a175c1dd8740fa402faaebe6a114fabec5aa4815d20b41ec651df6cb5ab880849b50509234cbae117a6e59eb77959515d01f9f943c70450183c2d72d3b9ec31e65db41fc90978b31f80537f8b77a5a119b83e8f5420837541402ba85d824325b3641f6d3de35c40ea246804db5a68b0d4897b5169488d76d0a32d2006cdf754e15bbfd4813501a2e88fa3beb1a4dcebf4ccf88d8ba9ea0a9f53355fdc8d44b77caca7c73ad939f9e57d3edabb3c3f60328fc2c01b41bcb4cb533a95477107f447e36a12456abf8c550d1ae359d0f7087a16d4528832a3120f28793d2383f2a2574853f19210895d39bf3840ef106729b7fbdff3ba208666c21b657c2fcdae9e9d801dc205922ee25edbdc6b789907b7fb49d41460184e9b570d3ae6c25ee801c2178ed6874c67901d7fbb73838d7433fa9ae7638ce007ed83fb517e5b0bb8b939b22097ff3db80dd647dbfbda0126e1a35270606e1b28936dfcc8decf4992448d567977bda52fbb4d4c147bb73c6aab1195d8f79e4e458141869569e9cee435140cff55224a94b204a4a103ddad3a8082a1665ad7c52cef851a9376efbf07f6310481a04ea49d0a58072f5089db9fdc54022c52b9321b2b10bd94ad1a739ba8fa4f54d3dade5a62c611bcb4362d7dfb8b0913df8d66021a5f2870aa75e334a72a8c77f8dfb0d5e88a5828017289ba62636dbaa95eebda33b79714e594f19dff834510366a89cab27f7d5af4cc496ae109d5eb907cb9386bd126706d963ed9e6d56439e8bbd735f506243aa2e38474cc53c153b37392e5e56b4c271d7182dd4d31395b3c76f0e6ec6fb79cf02498742c265daa98cccab211d0d1eef2e9e7009defd5c5b0fc2e9d6f3c1518fc964cdd98aa019b4b2a1db22f734788f6dcbe6e8b3fdeb9dfea69410c14c039ee6ec4489171f744e2af555773da6886a2a7a3ad5b032b85e91d56d2baf4176004a37a60e6a00b8dc08b99031d316766af541").unwrap();
+        let pt = Plaintext::from(Fp12Elem::decode(bytes).unwrap());
+        let src = &hex::decode("e8fe0292045bda2a9102749497e24e9502dd00d94670b6990e9aca44b7eed3416888bfba59e19bed709f7309a05e701b3a78dad9abd7764a6d5c4dfe")
+            .unwrap()[..];
+        let mut dest = [0u8; 60];
+        dest.copy_from_slice(src);
+        let expected_result = PrivateKey::new(dest);
+        assert_eq!(api.derive_private_key(&pt), expected_result);
+    }
+
     #[test]
     fn publickey_new_from_slice() {
-        let mut api = Api480::new();
+        let mut api = Recrypt480::new();
         let (_, pk1) = api.generate_key_pair().unwrap();
         let input = (pk1.x.0, pk1.y.0);
         let slice: (&[u8], &[u8]) = (&input.0, &input.1);
@@ -1469,15 +1490,31 @@ pub(crate) mod test {
         assert_eq!(pk_from_fixed.unwrap(), pk_from_slice.unwrap());
 
         assert_eq!(
-            ApiErr::InputWrongSize("PublicKey", 120),
+            RecryptErr::InputWrongSize("PublicKey", 120),
             PublicKey::new_from_slice((&input.0[..30], &input.1[..32])).unwrap_err()
+        )
+    }
+
+    #[test]
+    fn private_key_new_from_slice() {
+        let mut rand_bytes = DummyRandomBytes;
+        let input: [u8; 60] = rand_bytes.random_bytes_60();
+        let slice: &[u8] = &input;
+        let from_fixed = PrivateKey::new(input);
+        let from_slice = PrivateKey::new_from_slice(slice);
+
+        assert_eq!(from_fixed, from_slice.unwrap());
+
+        assert_eq!(
+            RecryptErr::InputWrongSize("PrivateKey", 60),
+            PrivateKey::new_from_slice(&input[..59]).unwrap_err()
         )
     }
 
     // note that this doesn't show that Drop is working properly, just that clear does
     #[test]
     fn private_key_clear() {
-        let (mut priv_key, _) = Api480::new().generate_key_pair().unwrap();
+        let (mut priv_key, _) = Recrypt480::new().generate_key_pair().unwrap();
         priv_key.clear();
         assert_eq!(SixtyBytes(priv_key.bytes().clone()), SixtyBytes([0u8; 60]));
         assert_eq!(priv_key._internal_key, Default::default())